refactor: move lsp hints to a separate module
